requirements.txt
# Core dependencies
google-api-python-client>=2.150.0
# Removed date version to allow 3.9 compatible releases
yt-dlp
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.1
# datetime is a standard library, no need to list here
pyjwt>=2.10.1
requests>=2.32.3
replicate>=1.0.4

# LLM
ollama>=0.4.0

# Video/Audio
moviepy==1.0.3
gtts>=2.5.4
pillow>=10.0.0
# Use headless version for Docker to save space and avoid UI library errors
opencv-python-headless>=4.8.0

# Utils
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
elevenlabs>=1.0.0

# FastAPI Stack
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.9
# Pydantic v2 is 3.9 compatible, but we cap it to ensure stability
pydantic>=2.0.0,<2.10.0
pydantic-settings>=2.0.0